ai-model/requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
pandas>=1.4.0
Flask>=2.2.0
pathlib2>=2.3.0
